Add buttons to create/join team onn dashboard

+++ src/controllers/applicationController.ts
@@ -129,7 +129,7 @@ export class ApplicationController implements ApplicationControllerInterface {
       return next(err);
     }
 
-    if (application.applicationStatus <= ApplicantStatus.Applied) {
+    if (application.applicationStatus <= ApplicantStatus.Applied && res.locals.applicationsOpen) {
       // Delete the application so they can re-apply
       try {
         await this._applicantService.remove(application.id);
